+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2017-10-XX
+### Added
+- Settings & account panels
+
+### Changed
+- Fix for sorting hide/show functionality
+
 ## [0.2.1] - 2017-10-13
 ### Added
 - Translation to German